Move 'convertToExportY' to shared
